+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 ### Bug fixes
 
+* Removing type checking in diversity calculations.  ([#1583](https://github.com/biocore/scikit-bio/issues/1583)).
+
 ### Deprecated functionality [stable]
 
 ### Deprecated functionality [experimental]