+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 ### Bug fixes
 
-* Removing type checking in diversity calculations.  ([#1583](https://github.com/biocore/scikit-bio/issues/1583)).
+* Relaxing type checking in diversity calculations.  ([#1583](https://github.com/biocore/scikit-bio/issues/1583)).
 
 ### Deprecated functionality [stable]